+++ extension.js
@@ -1,4 +1,4 @@
-const DEV_MODE = process.argv?.[2] == "-dev";
+const DEV_MODE = process.argv?.[2] === "-dev";
 const vscode = DEV_MODE ? null : require("vscode");
 const fs = require("fs");
 const packageJson = require("./package.json");
@@ -11,27 +11,50 @@ const LANGUAGE_SCOPE_PREFIX = "meta.embedded.inline";
 const REPOSITORY_SUFFIX = "block-scalar";
 const GLOBAL_STATE_VERSION = "version";
 
+/* eslint sort-keys: ["error", "asc"] */
 const LANGUAGES = {
+  bat: "source.batchfile",
+  bibtex: "text.bibtex",
   c: "source.c",
+  "c#": {
+    name: "csharp",
+    scopeName: "source.cs",
+  },
+  "c\\+\\+": {
+    name: "cpp",
+    scopeName: "source.cpp",
+  },
   clojure: "source.clojure",
   coffee: {
     name: "coffeescript",
     scopeName: "source.coffee",
   },
   cpp: "source.cpp",
-  "c\\+\\+": {
-    name: "cpp",
-    scopeName: "source.cpp",
-  },
-  csharp: "source.csharp",
+  csharp: "source.cs",
   css: "source.css",
-  diff: "source.diff",
+  cuda: {
+    name: "cuda-cpp",
+    scopeName: "source.cuda-cpp",
+  },
+  dart: "source.dart",
+  diff: {
+    scopeName: "source.diff",
+    stripIndent: true,
+  },
+  dockercompose: "source.yaml",
   dockerfile: "source.dockerfile",
-  dosbatch: "source.dosbatch",
+  "f#": {
+    name: "fsharp",
+    scopeName: "source.fsharp",
+  },
   fsharp: "source.fsharp",
   go: "source.go",
   groovy: "source.groovy",
+  handlebars: "text.html.handlebars",
+  hlsl: "source.hlsl",
   html: "text.html.derivative",
+  ini: "source.ini",
+  jade: "text.pug",
   java: "source.java",
   javascript: "source.js",
   js: {
@@ -39,24 +62,44 @@ const LANGUAGES = {
     scopeName: "source.js",
   },
   json: "source.json",
-  tex: "text.tex",
-  latex: "text.tex",
+  jsonc: "source.json.comments",
+  jsonl: "source.json.lines",
+  jsx: {
+    name: "javascriptreact",
+    scopeName: "source.js.jsx",
+  },
+  julia: "source.julia",
+  latex: {
+    scopeName: "text.tex.latex",
+    stripIndent: true,
+  },
+  less: "source.css.less",
+  log: "text.log",
   lua: "source.lua",
+  make: {
+    name: "makefile",
+    scopeName: "source.makefile",
+  },
   makefile: "source.makefile",
   markdown: {
     scopeName: "text.html.markdown",
     stripIndent: true,
   },
+  math: {
+    name: "markdown-math",
+    scopeName: "text.html.markdown.math",
+  },
   objc: {
     name: "objective-c",
     scopeName: "source.objc",
   },
-  perl: "source.perl",
-  pip: {
-    name: "pip-requirements",
-    scopeName: "source.pip-requirements",
+  objcpp: {
+    name: "objective-cpp",
+    scopeName: "source.objcpp",
   },
-  requirements: {
+  perl: "source.perl",
+  php: "text.html.php",
+  pip: {
     name: "pip-requirements",
     scopeName: "source.pip-requirements",
   },
@@ -65,14 +108,24 @@ const LANGUAGES = {
     scopeName: "source.js",
   },
   powershell: "source.powershell",
-  properties: "source.properties",
-  python: "source.python",
+  properties: "source.ini",
   py: {
     name: "python",
     scopeName: "source.python",
   },
+  python: "source.python",
   r: "source.r",
-  regex: "source.regexp.python",
+  raku: "source.perl.6",
+  razor: "text.html.cshtml",
+  regex: "source.js.regexp",
+  requirements: {
+    name: "pip-requirements",
+    scopeName: "source.pip-requirements",
+  },
+  rst: {
+    name: "restructuredtext",
+    scopeName: "source.rst",
+  },
   ruby: "source.ruby",
   rust: "source.rust",
   scss: "source.css.scss",
@@ -81,10 +134,9 @@ const LANGUAGES = {
     name: "shellscript",
     scopeName: "source.shell",
   },
-  slim: "source.slim",
   sql: "source.sql",
   swift: "source.swift",
-  typescript: "source.ts",
+  tex: "text.tex",
   ts: {
     name: "typescript",
     scopeName: "source.ts",
@@ -93,9 +145,13 @@ const LANGUAGES = {
     name: "typescriptreact",
     scopeName: "source.tsx",
   },
+  typescript: "source.ts",
+  vb: "source.asp.vb.net",
   xml: "text.xml",
+  xsl: "text.xml.xsl",
   yaml: "source.yaml",
 };
+/* eslint-disable sort-keys */
 
 const getEmbeddedLanguages = (languages) => {
   const ids = Object.keys(languages);
@@ -107,6 +163,7 @@ const getEmbeddedLanguages = (languages) => {
 const getPackageJson = (languages) => ({
   ...packageJson,
   contributes: {
+    ...packageJson.contributes,
     grammars: [
       {
         path: INJECTION_PATH,
@@ -115,30 +172,6 @@ const getPackageJson = (languages) => ({
         embeddedLanguages: getEmbeddedLanguages(languages),
       },
     ],
-    configuration: {
-      title: packageJson.displayName,
-      properties: {
-        [INCLUDE_CONFIG]: {
-          type: "object",
-          patternProperties: {
-            "^.*$": {
-              type: ["string", "object"],
-              properties: {
-                name: {
-                  type: "string",
-                },
-                scopeName: {
-                  type: "string",
-                },
-              },
-            },
-          },
-          default: {},
-          description:
-            "Use the key to define the language identifier with regex. Use the value to specify the language TextMate `scopeName`. By default the language identifier will be used as the language name. To change this, an object can be specified with the properties `name` and `scopeName`.",
-        },
-      },
-    },
   },
 });
 
@@ -155,7 +188,7 @@ const getRepository = (languages) => {
     entries.map(([id, { scopeName, stripIndent }]) => [
       `${id}-${REPOSITORY_SUFFIX}`,
       {
-        begin: `(?i)(?:(\\|)|(>))([1-9])?([-+])?[ \t]+(#[ \t]*(?:${id})[ \t]*\\n)`,
+        begin: `(?i)(?:(\\|)|(>))([1-9])?([-+])?\\s+(#\\s*(?:${id})\\s*\\n)`,
         beginCaptures: {
           1: {
             name: "keyword.control.flow.block-scalar.literal.yaml",
@@ -170,16 +203,7 @@ const getRepository = (languages) => {
             name: "storage.modifier.chomping-indicator.yaml",
           },
           5: {
-            patterns: [
-              {
-                begin: "#",
-                beginCaptures: {
-                  0: { name: "punctuation.definition.comment.yaml" },
-                },
-                end: "\\n",
-              },
-            ],
-            name: "comment.line.number-sign.yaml",
+            name: "entity.name.type.yaml",
           },
         },
         end: "^(?=\\S)|(?!\\G)",